Import material-ui/icons, add working sidenav from docs
Material-ui icons: https://material-ui.com/components/icons/

Docs on sidenav which uses material-ui icons:
https://material-ui.com/components/drawers/

The stock code for pages (ie an email system with inbox/drafts, etc)
was just stuffed in for the time being as pages for the real site
are yet to be discussed. This is more of a proof of concept that
material-ui is a solid choice for the project. I merely copy/pasted
the example code into the SideDrawer component, and in the infamous
words of a prolific developer:

"It just works" -Todd Howard