Utilize DayRecord CSS file using CSS modules
What wasn't clear from Max's course at a glance awhile later, is that
this project must be missing some CSS import. The course project
uses normall CSS files, but a guide I found (lniked below) said that
it is a crucial distinction to name files with "file.module.css" to
make a local scope. For now I am trying this.