Add props to day records for the `DayRecord CC`
I knew that props are React 101, but I was waiting for a fleshed out
parent/child relationship to relarn the basics. Super easy example and
more refactors related to props will follow soon.

ABOUT CC: First of all, the 'CC' is a new convention for shortening commit
messages, where 'CC' is 'Class Component'. React's way of handling
class components vs functional components can very greatly. I
feel like if there is space in a message it's worth mentioning the
type of component. Knowing what to expect going in since people may
not see the type in the diff or immediately recognize the type
based on the code structure is nice. I sure can't always understand
being this rusty myself. This is just a suggestion and may be cringy
in retrospect.