[Major] Add `Express.js` backend
Most file changes are from moving the frontend folders/files to the
root/s new `/client` folder. Server content is housed in the root's
`/server` folder and with its root level `package.json`. This
is explained further in `Documentation/Backend/Express_Server.md`.

`npm run dev` installs the packages for the client and server,
and then starts the server and client in that order at once.

Right now a vanilla HTML button on the home page pings the backend
and logs that it was hit in the devtools console.